Decouple ROS message types from the observation model (#27)
This patch decouples ROS message types from the likelihood field sensor model, allowing it to be moved to the main package. It also replaces the custom state class with SE(2) group elements using the sophus library. The only functional change with respect to the previous implementation is that we now support different orientations for the map origin.