Fix `perfect_odometry` rosbag
This patch fixes message timestamps in the rosbag to match the header
timestamps that were recorded using simulation time. The storage backend
was changed to mcap, the contents of the messages were not changed.
It also removes the last messages where there was a collision against one
of the walls.

Signed-off-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>