Change default AMCL params
This is a minor patch to make default params match the navigation AMCL
values in ROS 1 and ROS 2. This is convenient for users migrating from
the AMCL node provided by the navigation stack, if they define a
parameter file that does not mention all parameters, they can expect the
behavior to match that of the navigation node.

Signed-off-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>