Fix differential drive motion model (#267)
### Proposed changes

Going over the math a few times, it doesn't seem to me that the rotation
variance computation distributes over rotation composition. Not with
those nonlinear max, min, abs operations in between.

#### Type of change

- [x] 🐛 Bugfix (change which fixes an issue)
- [ ] 🚀 Feature (change which adds functionality)
- [ ] 📚 Documentation (change which fixes or extends documentation)

### Checklist

- [x] Lint and unit tests (if any) pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added necessary documentation (if appropriate)
- [x] All commits have been signed for
[DCO](https://developercertificate.org/)

### Additional comments

Not sure how to write a test for this. Would a formal proof (that
`rotation_variance(a) * rotation_variance(b) != rotation_variance(a *
b)`) do?

---------

Signed-off-by: Michel Hidalgo <michel@ekumenlabs.com>
Signed-off-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>
Co-authored-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>