Add assign action
View-action composition in the range library is limited (and often
produces the wrong result). This patch implements a new range adaptor
object that can be appended to any view closure to evaluate it
eagerly and assign the result to a given range, effectively converting
any view into an action.

It also adds an operator<<= overload to simplify syntax.

Signed-off-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>