Motion and sensor models are not mixins (#291)
Big patch to make motion and sensor models independent classes. It also
removes utilities that are no longer needed.

Related to #279.

💥 **Breaking change!** This removes some of the runtime support
utilities we added for mixins. Please note that statically defined
mixins are still fully supported (and will be until the migration is
complete) and runtime dispatch through virtual interfaces is still there
as well.

Signed-off-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>