Update system tests to use the new API
This patch removes system test dependency on beluga_amcl. It rewrites
the test with the new range-based API to run the cartesian product of
the selected particle filter variants, sensor models and motion models.

It is also meant to drive the development of the new API by highlighting
its current shortcomings.

Signed-off-by: Nahuel Espinosa <nespinosa@ekumenlabs.com>