Relocate `make_random_state()` method (#324)
### Proposed changes

Connected to #279. Follow-up to #322. This is a long due change, and as
I was reworking sensor models I figured it was appropriate. This patch
removes `make_random_state()` from sensor models. As we still support
mixins, it relocates `make_random_state()` to the `RandomStateGenerator`
mixin, which now takes both state and map as template parameters.

#### Type of change

- [ ] 🐛 Bugfix (change which fixes an issue)
- [x] 🚀 Feature (change which adds functionality)
- [ ] 📚 Documentation (change which fixes or extends documentation)

💥 **Breaking change!** Sensor models no longer have nor require a
`make_random_state()` method.

### Checklist

- [x] Lint and unit tests (if any) pass locally with my changes
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added necessary documentation (if appropriate)
- [x] All commits have been signed for
[DCO](https://developercertificate.org/)

### Additional comments

I'm not super fond of these changes to `RandomStateGenerator`, but they
are not unreasonable in context and it is arguably a temporary
workaround.

We also have to update https://github.com/Ekumen-OS/beluga-demos to
match.

Signed-off-by: Michel Hidalgo <michel@ekumenlabs.com>