Add Likelihood prob sensor model
Signed-off-by: fbattocchia <florencia.battocchia@creativa77.com.ar>

Changed the StateType of the likelihood field sensor model in
`beluga_vdb` to a template in order to be able to use it with 2D motion
models. And avoid issues like [#478 Unscented transform assertion when
using 2.5D covariance
matrix](https://github.com/Ekumen-OS/beluga/issues/478)

- [ ] 🐛 Bugfix (change which fixes an issue)
- [X] 🚀 Feature (change which adds functionality)
- [ ] 📚 Documentation (change which fixes or extends documentation)

- [X] Lint and unit tests (if any) pass locally with my changes
- [X] I have added tests that prove my fix is effective or that my
feature works
- [X] I have added necessary documentation (if appropriate)
- [X] All commits have been signed for
[DCO](https://developercertificate.org/)

I changed the name of the sensor from `LikelihoodFieldModel3` to
`VDBLikelihoodFieldModel` because I think it might be confusing, since
the model will accept 2D particles.

---------

Signed-off-by: Pablo Vela <pablovela@naver.com>