Add likelihood prob sensor model and enable Beluga's flavour of clustering (#476)
### Proposed changes

The `LikelihoodFieldProb` sensor model is incorporated, based on the
implementation presented in the book Probabilistic Robotics. To achieve
this, the `LikelihoodFieldModelCommon` class is created, which
implements the Likelihood Field shared by both the `LikelihoodFieldProb`
and `LikelihoodField` models.

The `LikelihoodFieldProb` model, in addition to following the book's
implementation, includes the summation of the logarithm of probabilities
to prevent underflow when multiplying numerous small probabilities.
Additionally, the Cluster-Based Estimation algorithm is integrated,
enhancing the model's robustness against unmapped objects.

Furthermore, unit tests are implemented for the `LikelihoodFieldProb`
sensor model.

#### Type of change

- [ ] 🐛 Bugfix (change which fixes an issue)
- [x] 🚀 Feature (change which adds functionality)
- [ ] 📚 Documentation (change which fixes or extends documentation)

### Checklist

_Put an `x` in the boxes that apply. This is simply a reminder of what
we will require before merging your code._

- [x] Lint and unit tests (if any) pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] All commits have been signed for
[DCO](https://developercertificate.org/)

Signed-off-by: Pablo Vela <pablovela@naver.com>